requirements.txt

pandas
tinytag
tk
Pillow
python-pptx
unidecode
num2words
pygame
mutagen
ftfy